refactor(test): audit truncation.rs; rename to partial.rs
- Rename truncation.rs → partial.rs with module doc explaining partial
  change proofs (paginated sync via limit parameter)
- Add parameterized test_partial_proof_verified covering 8 edge cases:
  truncation, limit=count, limit>count, empty results, bounded+limited
- Clean up existing tests: rename db_a/db_b → source/target, remove
  dead root1_a variable, add block comments
- Move test_spurious_{put,delete}_at_start_key_boundary → attack.rs
- Move test_generator_uses_end_key_for_complete_proof → empty_and_bounded.rs
- Move test_disjoint_proof → edge_cases.rs
