+++ firewood/src/merkle/tests/change/mod.rs
@@ -82,8 +82,8 @@ mod divergent_child;
 mod edge_cases;
 mod empty_and_bounded;
 mod fuzz;
+mod partial;
 mod pipeline;
 mod regression;
 mod root_hash;
 mod structural;
-mod truncation;
